Revert to rmdir capability that was erroneously removed

+++ pypiper/pypiper.py
@@ -806,7 +806,8 @@ class PipelineManager(object):
 				files = glob.glob(regex)
 				for file in files:
 					with open(self.cleanup_file, "a") as myfile:
-						myfile.write("rm " + file + "\n")
+						if os.path.isfile(file): myfile.write("rm " + file + "\n")
+						elif os.path.isdir(file): myfile.write("rmdir " + file + "\n")
 			except:
 				pass
 		elif conditional:
@@ -834,7 +835,10 @@ class PipelineManager(object):
 					# and delete the files
 					for file in files:
 						print("`rm " + file + "`")
-						os.remove(os.path.join(file))
+							os.remove(os.path.join(file))
+						elif os.path.isdir(file):
+							print("`rmdir " + file + "`")
+							os.rmdir(os.path.join(file))
 				except:
 					pass
 
@@ -849,8 +853,12 @@ class PipelineManager(object):
 						files = glob.glob(expr)
 						while files in self.cleanup_list_conditional: self.cleanup_list_conditional.remove(files)
 						for file in files:
-							print("rm " + file)
-							os.remove(os.path.join(file))
+							if os.path.isfile(file):
+								print("rm " + file)
+								os.remove(os.path.join(file))
+							elif os.path.isdir(file):
+								print("`rmdir " + file + "`")
+								os.rmdir(os.path.join(file))
 					except:
 						pass
 			else:
@@ -862,7 +870,8 @@ class PipelineManager(object):
 						files = glob.glob(expr)
 						for file in files:
 							with open(self.cleanup_file, "a") as myfile:
-								myfile.write("`rm " + file + "`\n")
+								if os.path.isfile(file): myfile.write("`rm " + file + "`\n")
+								elif os.path.isdir(file): myfile.write("`rmdir " + file + "`\n")
 					except:
 						pass
